Split google and facebook in two different services
+ minor fix
